TREX: TRree sap flow EXtractor.

Performs data assimilation, processing and analyses on sap flow data obtained with the thermal dissipation method (TDM). The package includes functions for gap filling time-series data, detecting outliers, calculating data-processing uncertainties and, generating uniform data output and visualisation. The package is designed to deal with large quantities of data and apply commonly used data-processing methods. The functions have been validated on data collected from different tree species across the northern hemisphere (Peters et al. 2018 <doi: 10.1111/nph.15241>).

See also

Author

Maintainer: Richard Peters richardlouispeters3@hotmail.com (ORCID) [copyright holder]

Authors:

Christoforos Pappas (ORCID) [contributor]

Alexander Hurley (ORCID) [contributor]

Other contributors:

Rafael Poyatos [contributor]

Victor Flo [contributor]

Roman Zweifel [contributor]

Willem Goossens [contributor]

Kathy Steppe [thesis advisor, etcher, funder]