Calibration Measurements

Returns raw calibration experiment data obtained from literature, with K values combined with gravimetrically determined sap flux density, as detailed in Flo et al. (2019). The data.frame contains 22 studies with 37 different species. The data is used within the tdm_cal.sfd function to calculate sap flux density. Description on the genus, species, calibration material, wood porosity and diameter of the stem is provided in Flo et al. (2019). The presented data is open for public use.

cal.data

Format

Provides a data.frame with 4024 rows and 10 columns.

Study

Study from which the data originates (see Flo et al. 2019) (character)

Method

Heat-based sap flow measurement method (TD = Thermal Dissipation) (character)

Genus

Monitored genus (character)

Species

Monitored species (character)

Calibration.material

Description on the calibration method that was used, including stem segment, whole plant and whole plant without roots (character)

Wood.porosity

Wood structure type of the examined species, including coniferous, diffuse-porous, ring-porous and monocots (character)

Diameter

Diameter at breast height of the calibration subject (in cm) (numeric)

k

Proportional difference between \(\Delta T\) and \(\Delta T_{max}\) measured by the thermal dissipation probes (unitless; numeric)

SFD

Sap flux density measured gravimetrically (in \(cm^3 cm^{-2} h^{-1}\); numeric)

Granier

Sap flux density calculated according to Granier et al. 1985 using k (in \(cm^3 cm^{-2} h^{-1}\); using \(43.84 k^{1.231}\)) (numeric)

Details

Currently included studies are given below; individual labels (quoted) can be applied in tdm_cal.sfd (argument "study"):

Braun and Schmid 1999

Cabibel et al. 1991

Cain 2009

Chan 2015

Fuchs et al. 2017

Granier 1985

Gutierrez & Santiago 2005

Herbst et al. 2007

Liu et al. 2008

Lu 2002

Lu and Chacko 1998

Oliveira et al. 2006

Paudel et al. 2013

Rubilar et al. 2016

Schmidt-walter et al. 2014

Sperling et al. 2012

Sugiura et al. 2009

Sun et al. 2012

Vellame et al. 2009

Hubbard et al. 2010

Peters et al. 2017

Steppe et al. 2010

References

Flo V, Martinez-Vilalta J, Steppe K, Schuldt B, Poyatos, R. 2019. A synthesis of bias and uncertainty in sap flow methods. Agricultural and Forest Meteorology 271:362-374. doi: 10.1016/j.agrformet.2019.03.012

Granier A. 1985. Une nouvelle methode pour la measure du flux de seve brute dans le tronc des arbres. Annales des Sciences Forestieres 42:193–200. doi: 10.1051/forest:19850204