Predict allometric parameters from DBH

Predict allometric parameters from DBH

predict_allometry(
  input_dbh_cm,
  parameter = c("Bark_cm", "Sap_cm", "Saparea_cm2"),
  species = NULL,
  ...
)

Arguments

input_dbh_cm

numeric, single value or vector of DBH measurements in cm

parameter

character, one or multiple of "Bark_cm", "Sap_cm", "Saparea_cm2"

species

character, one or multiple of the species given in allometry.data. Defaults to NULL, selecting all species.

...

additional arguments passed to predict()

Value

A data.frame containing predictions for each provided parameter and species

Examples

  predict_allometry(input_dbh_cm = 10:15)
#>       parameter               species       value
#> 1       Bark_cm        Abies balsamea   0.1828331
#> 2       Bark_cm        Abies balsamea   0.1994880
#> 3       Bark_cm        Abies balsamea   0.2161429
#> 4       Bark_cm        Abies balsamea   0.2327977
#> 5       Bark_cm        Abies balsamea   0.2494526
#> 6       Bark_cm        Abies balsamea   0.2661074
#> 7       Bark_cm        Acer saccharum   0.3894793
#> 8       Bark_cm        Acer saccharum   0.4195945
#> 9       Bark_cm        Acer saccharum   0.4497098
#> 10      Bark_cm        Acer saccharum   0.4798250
#> 11      Bark_cm        Acer saccharum   0.5099402
#> 12      Bark_cm        Acer saccharum   0.5400554
#> 13      Bark_cm          Picea glauca   0.2056120
#> 14      Bark_cm          Picea glauca   0.2214723
#> 15      Bark_cm          Picea glauca   0.2373326
#> 16      Bark_cm          Picea glauca   0.2531929
#> 17      Bark_cm          Picea glauca   0.2690532
#> 18      Bark_cm          Picea glauca   0.2849135
#> 19      Bark_cm   Populus tremuloides   0.2097614
#> 20      Bark_cm   Populus tremuloides   0.2285580
#> 21      Bark_cm   Populus tremuloides   0.2473545
#> 22      Bark_cm   Populus tremuloides   0.2661510
#> 23      Bark_cm   Populus tremuloides   0.2849476
#> 24      Bark_cm   Populus tremuloides   0.3037441
#> 25      Bark_cm    Thuja occidentalis   0.2184624
#> 26      Bark_cm    Thuja occidentalis   0.2271448
#> 27      Bark_cm    Thuja occidentalis   0.2358273
#> 28      Bark_cm    Thuja occidentalis   0.2445097
#> 29      Bark_cm    Thuja occidentalis   0.2531921
#> 30      Bark_cm    Thuja occidentalis   0.2618746
#> 31      Bark_cm        Fraxinus nigra   0.2081129
#> 32      Bark_cm        Fraxinus nigra   0.2244296
#> 33      Bark_cm        Fraxinus nigra   0.2407463
#> 34      Bark_cm        Fraxinus nigra   0.2570630
#> 35      Bark_cm        Fraxinus nigra   0.2733796
#> 36      Bark_cm        Fraxinus nigra   0.2896963
#> 37      Bark_cm     Fagus grandifolia   0.1809484
#> 38      Bark_cm     Fagus grandifolia   0.1879238
#> 39      Bark_cm     Fagus grandifolia   0.1948992
#> 40      Bark_cm     Fagus grandifolia   0.2018746
#> 41      Bark_cm     Fagus grandifolia   0.2088500
#> 42      Bark_cm     Fagus grandifolia   0.2158254
#> 43      Bark_cm           Acer rubrum   0.2984239
#> 44      Bark_cm           Acer rubrum   0.3048401
#> 45      Bark_cm           Acer rubrum   0.3112563
#> 46      Bark_cm           Acer rubrum   0.3176725
#> 47      Bark_cm           Acer rubrum   0.3240887
#> 48      Bark_cm           Acer rubrum   0.3305049
#> 49      Bark_cm     Betula papyrifera   0.2516888
#> 50      Bark_cm     Betula papyrifera   0.2685465
#> 51      Bark_cm     Betula papyrifera   0.2854042
#> 52      Bark_cm     Betula papyrifera   0.3022619
#> 53      Bark_cm     Betula papyrifera   0.3191196
#> 54      Bark_cm     Betula papyrifera   0.3359772
#> 55      Bark_cm Betula alleghaniensis   0.2369803
#> 56      Bark_cm Betula alleghaniensis   0.2586018
#> 57      Bark_cm Betula alleghaniensis   0.2802233
#> 58      Bark_cm Betula alleghaniensis   0.3018448
#> 59      Bark_cm Betula alleghaniensis   0.3234664
#> 60      Bark_cm Betula alleghaniensis   0.3450879
#> 61      Bark_cm         Quercus rubra   0.6378782
#> 62      Bark_cm         Quercus rubra   0.6444177
#> 63      Bark_cm         Quercus rubra   0.6509572
#> 64      Bark_cm         Quercus rubra   0.6574966
#> 65      Bark_cm         Quercus rubra   0.6640361
#> 66      Bark_cm         Quercus rubra   0.6705756
#> 67      Bark_cm          Picea rubens   0.2212738
#> 68      Bark_cm          Picea rubens   0.2350518
#> 69      Bark_cm          Picea rubens   0.2488298
#> 70      Bark_cm          Picea rubens   0.2626078
#> 71      Bark_cm          Picea rubens   0.2763858
#> 72      Bark_cm          Picea rubens   0.2901638
#> 73      Bark_cm         Picea mariana   0.1920401
#> 74      Bark_cm         Picea mariana   0.1923496
#> 75      Bark_cm         Picea mariana   0.1926591
#> 76      Bark_cm         Picea mariana   0.1929686
#> 77      Bark_cm         Picea mariana   0.1932780
#> 78      Bark_cm         Picea mariana   0.1935875
#> 79      Bark_cm       Pinus banksiana   0.2746403
#> 80      Bark_cm       Pinus banksiana   0.2785581
#> 81      Bark_cm       Pinus banksiana   0.2824759
#> 82      Bark_cm       Pinus banksiana   0.2863937
#> 83      Bark_cm       Pinus banksiana   0.2903115
#> 84      Bark_cm       Pinus banksiana   0.2942293
#> 85       Sap_cm        Abies balsamea   2.3274513
#> 86       Sap_cm        Abies balsamea   2.4927363
#> 87       Sap_cm        Abies balsamea   2.6580213
#> 88       Sap_cm        Abies balsamea   2.8233063
#> 89       Sap_cm        Abies balsamea   2.9885913
#> 90       Sap_cm        Abies balsamea   3.1538764
#> 91       Sap_cm        Acer saccharum   2.9492098
#> 92       Sap_cm        Acer saccharum   3.1031662
#> 93       Sap_cm        Acer saccharum   3.2571226
#> 94       Sap_cm        Acer saccharum   3.4110789
#> 95       Sap_cm        Acer saccharum   3.5650353
#> 96       Sap_cm        Acer saccharum   3.7189917
#> 97       Sap_cm          Picea glauca   2.0609779
#> 98       Sap_cm          Picea glauca   2.1610736
#> 99       Sap_cm          Picea glauca   2.2611693
#> 100      Sap_cm          Picea glauca   2.3612650
#> 101      Sap_cm          Picea glauca   2.4613607
#> 102      Sap_cm          Picea glauca   2.5614564
#> 103      Sap_cm   Populus tremuloides   1.8988606
#> 104      Sap_cm   Populus tremuloides   2.3835701
#> 105      Sap_cm   Populus tremuloides   2.8682796
#> 106      Sap_cm   Populus tremuloides   3.3529891
#> 107      Sap_cm   Populus tremuloides   3.8376986
#> 108      Sap_cm   Populus tremuloides   4.3224081
#> 109      Sap_cm    Thuja occidentalis   1.4165788
#> 110      Sap_cm    Thuja occidentalis   1.4389873
#> 111      Sap_cm    Thuja occidentalis   1.4613958
#> 112      Sap_cm    Thuja occidentalis   1.4838044
#> 113      Sap_cm    Thuja occidentalis   1.5062129
#> 114      Sap_cm    Thuja occidentalis   1.5286215
#> 115      Sap_cm        Fraxinus nigra   2.9689847
#> 116      Sap_cm        Fraxinus nigra   3.4922694
#> 117      Sap_cm        Fraxinus nigra   4.0155541
#> 118      Sap_cm        Fraxinus nigra   4.5388388
#> 119      Sap_cm        Fraxinus nigra   5.0621235
#> 120      Sap_cm        Fraxinus nigra   5.5854082
#> 121      Sap_cm     Fagus grandifolia   2.6007336
#> 122      Sap_cm     Fagus grandifolia   2.8498929
#> 123      Sap_cm     Fagus grandifolia   3.0990523
#> 124      Sap_cm     Fagus grandifolia   3.3482116
#> 125      Sap_cm     Fagus grandifolia   3.5973709
#> 126      Sap_cm     Fagus grandifolia   3.8465302
#> 127      Sap_cm           Acer rubrum   2.2609850
#> 128      Sap_cm           Acer rubrum   2.4310307
#> 129      Sap_cm           Acer rubrum   2.6010764
#> 130      Sap_cm           Acer rubrum   2.7711221
#> 131      Sap_cm           Acer rubrum   2.9411678
#> 132      Sap_cm           Acer rubrum   3.1112135
#> 133      Sap_cm     Betula papyrifera   3.1132646
#> 134      Sap_cm     Betula papyrifera   3.4245889
#> 135      Sap_cm     Betula papyrifera   3.7359133
#> 136      Sap_cm     Betula papyrifera   4.0472376
#> 137      Sap_cm     Betula papyrifera   4.3585620
#> 138      Sap_cm     Betula papyrifera   4.6698864
#> 139      Sap_cm Betula alleghaniensis   3.6931406
#> 140      Sap_cm Betula alleghaniensis   3.8902101
#> 141      Sap_cm Betula alleghaniensis   4.0872796
#> 142      Sap_cm Betula alleghaniensis   4.2843490
#> 143      Sap_cm Betula alleghaniensis   4.4814185
#> 144      Sap_cm Betula alleghaniensis   4.6784879
#> 145      Sap_cm         Quercus rubra   1.7137171
#> 146      Sap_cm         Quercus rubra   1.9242018
#> 147      Sap_cm         Quercus rubra   2.1346866
#> 148      Sap_cm         Quercus rubra   2.3451714
#> 149      Sap_cm         Quercus rubra   2.5556561
#> 150      Sap_cm         Quercus rubra   2.7661409
#> 151      Sap_cm          Picea rubens   3.8141896
#> 152      Sap_cm          Picea rubens   4.2456251
#> 153      Sap_cm          Picea rubens   4.6770605
#> 154      Sap_cm          Picea rubens   5.1084959
#> 155      Sap_cm          Picea rubens   5.5399313
#> 156      Sap_cm          Picea rubens   5.9713667
#> 157      Sap_cm         Picea mariana   1.1252255
#> 158      Sap_cm         Picea mariana   1.1934694
#> 159      Sap_cm         Picea mariana   1.2617133
#> 160      Sap_cm         Picea mariana   1.3299572
#> 161      Sap_cm         Picea mariana   1.3982011
#> 162      Sap_cm         Picea mariana   1.4664450
#> 163      Sap_cm       Pinus banksiana   1.2168844
#> 164      Sap_cm       Pinus banksiana   1.3269039
#> 165      Sap_cm       Pinus banksiana   1.4369234
#> 166      Sap_cm       Pinus banksiana   1.5469429
#> 167      Sap_cm       Pinus banksiana   1.6569624
#> 168      Sap_cm       Pinus banksiana   1.7669819
#> 169 Saparea_cm2        Abies balsamea  48.1131203
#> 170 Saparea_cm2        Abies balsamea  57.5684234
#> 171 Saparea_cm2        Abies balsamea  67.8141722
#> 172 Saparea_cm2        Abies balsamea  78.8423135
#> 173 Saparea_cm2        Abies balsamea  90.6455134
#> 174 Saparea_cm2        Abies balsamea 103.2170439
#> 175 Saparea_cm2        Acer saccharum  64.0875331
#> 176 Saparea_cm2        Acer saccharum  74.2661563
#> 177 Saparea_cm2        Acer saccharum  84.9638625
#> 178 Saparea_cm2        Acer saccharum  96.1605353
#> 179 Saparea_cm2        Acer saccharum 107.8383634
#> 180 Saparea_cm2        Acer saccharum 119.9814270
#> 181 Saparea_cm2          Picea glauca  53.1810477
#> 182 Saparea_cm2          Picea glauca  61.2890050
#> 183 Saparea_cm2          Picea glauca  69.7657729
#> 184 Saparea_cm2          Picea glauca  78.5952764
#> 185 Saparea_cm2          Picea glauca  87.7633511
#> 186 Saparea_cm2          Picea glauca  97.2573937
#> 187 Saparea_cm2   Populus tremuloides  51.6101853
#> 188 Saparea_cm2   Populus tremuloides  64.1426450
#> 189 Saparea_cm2   Populus tremuloides  78.2235934
#> 190 Saparea_cm2   Populus tremuloides  93.8913758
#> 191 Saparea_cm2   Populus tremuloides 111.1820992
#> 192 Saparea_cm2   Populus tremuloides 130.1299217
#> 193 Saparea_cm2    Thuja occidentalis  39.0959932
#> 194 Saparea_cm2    Thuja occidentalis  44.6031181
#> 195 Saparea_cm2    Thuja occidentalis  50.3054258
#> 196 Saparea_cm2    Thuja occidentalis  56.1926883
#> 197 Saparea_cm2    Thuja occidentalis  62.2559738
#> 198 Saparea_cm2    Thuja occidentalis  68.4873995
#> 199 Saparea_cm2        Fraxinus nigra  58.6582576
#> 200 Saparea_cm2        Fraxinus nigra  72.3868833
#> 201 Saparea_cm2        Fraxinus nigra  87.7078151
#> 202 Saparea_cm2        Fraxinus nigra 104.6499435
#> 203 Saparea_cm2        Fraxinus nigra 123.2403037
#> 204 Saparea_cm2        Fraxinus nigra 143.5043251
#> 205 Saparea_cm2     Fagus grandifolia  55.9730320
#> 206 Saparea_cm2     Fagus grandifolia  67.8216911
#> 207 Saparea_cm2     Fagus grandifolia  80.8160315
#> 208 Saparea_cm2     Fagus grandifolia  94.9575115
#> 209 Saparea_cm2     Fagus grandifolia 110.2474742
#> 210 Saparea_cm2     Fagus grandifolia 126.6871642
#> 211 Saparea_cm2           Acer rubrum  44.5619929
#> 212 Saparea_cm2           Acer rubrum  54.0627206
#> 213 Saparea_cm2           Acer rubrum  64.4945459
#> 214 Saparea_cm2           Acer rubrum  75.8597209
#> 215 Saparea_cm2           Acer rubrum  88.1603222
#> 216 Saparea_cm2           Acer rubrum 101.3982757
#> 217 Saparea_cm2     Betula papyrifera  56.8949703
#> 218 Saparea_cm2     Betula papyrifera  69.2851276
#> 219 Saparea_cm2     Betula papyrifera  82.9384267
#> 220 Saparea_cm2     Betula papyrifera  97.8622816
#> 221 Saparea_cm2     Betula papyrifera 114.0635497
#> 222 Saparea_cm2     Betula papyrifera 131.5486121
#> 223 Saparea_cm2 Betula alleghaniensis  81.9032781
#> 224 Saparea_cm2 Betula alleghaniensis  95.4974234
#> 225 Saparea_cm2 Betula alleghaniensis 109.8690050
#> 226 Saparea_cm2 Betula alleghaniensis 124.9921150
#> 227 Saparea_cm2 Betula alleghaniensis 140.8436895
#> 228 Saparea_cm2 Betula alleghaniensis 157.4030107
#> 229 Saparea_cm2         Quercus rubra  42.9828281
#> 230 Saparea_cm2         Quercus rubra  52.0903537
#> 231 Saparea_cm2         Quercus rubra  62.0801037
#> 232 Saparea_cm2         Quercus rubra  72.9533361
#> 233 Saparea_cm2         Quercus rubra  84.7112093
#> 234 Saparea_cm2         Quercus rubra  97.3547972
#> 235 Saparea_cm2          Picea rubens  69.7871085
#> 236 Saparea_cm2          Picea rubens  84.6105623
#> 237 Saparea_cm2          Picea rubens 100.8766101
#> 238 Saparea_cm2          Picea rubens 118.5878774
#> 239 Saparea_cm2          Picea rubens 137.7467835
#> 240 Saparea_cm2          Picea rubens 158.3555716
#> 241 Saparea_cm2         Picea mariana  32.4347012
#> 242 Saparea_cm2         Picea mariana  38.4441731
#> 243 Saparea_cm2         Picea mariana  44.8976357
#> 244 Saparea_cm2         Picea mariana  51.7867879
#> 245 Saparea_cm2         Picea mariana  59.1041322
#> 246 Saparea_cm2         Picea mariana  66.8428433
#> 247 Saparea_cm2       Pinus banksiana  35.0465965
#> 248 Saparea_cm2       Pinus banksiana  42.0880504
#> 249 Saparea_cm2       Pinus banksiana  49.7448877
#> 250 Saparea_cm2       Pinus banksiana  58.0128840
#> 251 Saparea_cm2       Pinus banksiana  66.8881795
#> 252 Saparea_cm2       Pinus banksiana  76.3672233

  predict_allometry(input_dbh_cm = 10:15,
    parameter = "Bark_cm",
    species   = c("Pinus banksiana",
                "Betula alleghaniensis"))
#>    parameter               species     value
#> 1    Bark_cm       Pinus banksiana 0.2746403
#> 2    Bark_cm       Pinus banksiana 0.2785581
#> 3    Bark_cm       Pinus banksiana 0.2824759
#> 4    Bark_cm       Pinus banksiana 0.2863937
#> 5    Bark_cm       Pinus banksiana 0.2903115
#> 6    Bark_cm       Pinus banksiana 0.2942293
#> 7    Bark_cm Betula alleghaniensis 0.2369803
#> 8    Bark_cm Betula alleghaniensis 0.2586018
#> 9    Bark_cm Betula alleghaniensis 0.2802233
#> 10   Bark_cm Betula alleghaniensis 0.3018448
#> 11   Bark_cm Betula alleghaniensis 0.3234664
#> 12   Bark_cm Betula alleghaniensis 0.3450879