Heartwood correction

The function corrects for the proportion of the probe that is installed within the non-conductive heartwood according to Clearwater et al. (1999). The function requires \(\Delta T_{max}\), the probe length and the sapwood thickness. The correction is applied on the \(\Delta T\) (or \(\Delta V\)) values and \(K\) is recalculated accordingly. When an is.trex-compliant object is provided, the \(K\) values for each method are determined (see tdm_dt.max.

tdm_hw.cor (input, dt.max, probe.length = 20,
            sapwood.thickness = 18, df = FALSE)

Arguments

input

A tdm_dt.max ouput or is.trex-compliant object of \(\Delta T \) (or \(\Delta V\)) values containing a timestamp and a value column.

dt.max

Optional zoo object or data.frame (columns = “timestamp” or “value”) containing the \(\Delta T_{max}\) when no is.trex-compliant object is provided.

probe.length

Numeric, the length of the TDM probes in mm.

sapwood.thickness

Numeric, the sapwood thickness in mm.

df

Logical; If TRUE, output is provided in a data.frame format with a timestamp and a value column. If FALSE, output is provided as a zoo vector object (default = FALSE).

Value

A zoo object or data.frame in the appropriate format for other functionalities. See tdm_dt.max for output specifications. All \(K\) values for each method are provided when an output from tdm_dt.max was provided. If individual time series are provided for input and tdm_dt.max an alternative output is provided:

input

= \(\Delta T\) input data.

dt.max

\(\Delta T_{max}{\Delta Tmax}\) input data.

dtsw

Corrected \(\Delta T\) data.

k.value

\(K\) values calculated according to Clearwater et al. (1999).

settings

data.frame of the applied probe.length and sapwood.thickness

Details

The function applied the correction provided by Clearwater et al. 1999. \(\Delta T\) (or \(\Delta V\)) was corrected (denoted as \(\Delta T_{sw}\)) for the proportion of the probe that was inserted into the conducting sapwood vs the proportion of the probe that was inserted into the nonconductive heartwood (\(\gamma\) in mm mm-1). Together with \(\Delta T_{max}\), \(\Delta T\) was corrected according to the following equation: $$\Delta T_{sw} = (\Delta T~-~(1~–~\gamma) * \Delta T_{max}) / \gamma$$

\(\Delta T_{sw}\) together with \(\Delta T_{max}\) is then recalculated to \(K\).

References

Clearwater MJ, Meinzer FC, Andrade JL, Goldstein G, Holbrook NM. 1999. Potential errors in measurement of nonuniform sap flow using heat dissipation probes. Tree Physiology 19:681–687 doi:10.1093/treephys/19.10.681

Examples

if (FALSE) {
#correct for probes being inserted into the heartwood
raw   <-is.trex(example.data(type="doy"),
          tz="GMT",time.format="%H:%M",solar.time=TRUE,
          long.deg=7.7459,ref.add=FALSE)
input <- dt.steps(input=raw,
                   start="2014-05-08 00:00",
                   end="2014-07-25 00:50",
                  time.int=15,max.gap=60,decimals=6,df=F)
input[which(input<0.2)]<-NA
input <-tdm_dt.max(input, methods=c("pd","mw","dr"),
                 det.pd=TRUE,interpolate=FALSE,max.days=10,df=FALSE)
output.data<-tdm_hw.cor(input,probe.length=20,
                   sapwood.thickness=18,df=FALSE)
plot(output.data$k.dr,col="orange")
lines(input$k.dr)
}