Generate example TDM input data

This function returns a data.frame containing standard TDM (thermal dissipation method) measurements provided in two different formats. The data is obtained from tdm.data where \(\Delta V\) measurements are given for Norway spruce (Picea abies Karts.) growing in a valley in the Swiss Alps. See tdm.data for additional details.

example.data(type = "timestamp")

Arguments

type
Character string, indicating whether the example data should be displayed with a timestamp (default = “timestamp”) or whether a separate day of day of year and hour column are provided (“doy”).

Value

A data.frame containing TDM measurements according to a specific type.

Details

This dataset can be applied for testing the functions provided in TREX.

Examples

# get example data
input_data <- example.data(type = "timestamp")
input_data <- example.data(type = "doy")
head(input_data)

#>   year doy  hour value
#> 1 2012 108 15:00 0.444
#> 2 2012 108 15:15 0.541
#> 3 2012 108 15:30 0.560
#> 4 2012 108 15:45 0.568
#> 5 2012 108 16:00 0.572
#> 6 2012 108 16:15 0.545