Data cleaning and outlier detection

This function launches a Shiny application that (1) visualizes raw and outlier-free time series interactively (using plotly), (2) highlights automatically detected outliers, (3) allows the user to revise the automatically detected outliers and manually include data points, and (4) exports the original data and the outlier-free time series in an is.trex-compliant object that can be further processed.

outlier()

Value

Once the application is launched, the user can load an .RData file where a data.frame with the timestamp and sensor data are stored. The timestamp in this data.frame should be of class Date. Many columns can be included (corresponding, for example, to different sensors) and the user can select through the application the x and y axes of the interactive time series plots. In addition, the user can provide the units of the imported data (e.g., degrees \(C\) or \(mV\) for \(\Delta T\) or \(\Delta V\), respectively). A parameter (alpha) for automatic outlier detection can be supplied. More specifically, the automatic identification of outliers is based on a two-step procedure: i) the Tukey’s method (Tukey, 1977) is applied to detect statistical outliers as values falling outside the range \([q_{0.25} - alpha * IQR, q_{0.75} + alpha * IQR]\), where \(IQR\) is the interquartile range (\(q_{0.75} - q_{0.25}\)) with \(q_{0.25}\) denoting the 25% lower quartile and \(q_{0.75}\) the 75% upper quartile, and alpha is a user-defined parameter (default value alpha = 3; although visual inspection through the interactive plots allows for adjusting alpha and optimizing the automatic detection of outliers), and ii) the first difference (or lag-1 differences) of the raw data are calculated and data points with lag-1 differences greater than the mean of the raw input time series, are excluded.

The function does not return a value, but allows the user to write the raw and outlier-free data in a .RData file to disk, for subsequent import via load().

Examples

if (FALSE) {
# find example file path
system.file("exdata", "example.RData", package = "TREX", mustWork = TRUE)
# either copy-paste this into the navigation bar of the file selection window
# or navigate here manually for selection

# launch shiny application
outlier()

}